Remove old carousel code
