+++ components/content/carousel/MediaViewerModal/ImageView.tsx
