+++ components/content/carousel/ThumbnailList.tsx
