+++ package.json
@@ -111,7 +111,6 @@
     "react-native": "0.76.9",
     "react-native-collapsible": "^1.6.0",
     "react-native-gesture-handler": "~2.20.2",
-    "react-native-image-viewing": "^0.2.2",
     "react-native-logs": "^5.0.1",
     "react-native-maps": "1.18.0",
     "react-native-pager-view": "6.5.1",
@@ -174,7 +173,6 @@
     "doctor": {
       "reactNativeDirectoryCheck": {
         "exclude": [
-          "/react-native-image-viewing/",
           "/@babel/plugin-proposal-nullish-coalescing-operator/",
           "/@babel/plugin-proposal-optional-chaining/",
           "/@babel/plugin-transform-arrow-functions/",
